split test suites to ensure they run
